remove contentcodable
