+++ Sources/HTTP/Content/Message+MediaType.swift
@@ -1,0 +1,17 @@
+// MARK: Message
+
+extension Message {
+    /// The MediaType inside the `Message` `Headers`' "Content-Type"
+    public var mediaType: MediaType? {
+        get {
+            guard let contentType = headers[.contentType] else {
+                return nil
+            }
+            
+            return MediaType(string: contentType)
+        }
+        set {
+            headers[.contentType] = newValue?.description
+        }
+    }
+}
